Move PipelineAbortHelper to hudsontrigger package
Relocate PipelineAbortHelper (and its test) out of coordination.hazelcast
into hudsontrigger, widening it from package-private to public so
HazelcastBuildMemoryStorage can use it across the package boundary.
